refactor initial alpaka  vendor interface
- rename `alpaak::vendor` to `alpaka::fn`
- update documantation
- add generic fallback policy `toGeneric`
